backend/requirements.txt
python-dotenv
langchain
langchain_community
google-generativeai
langchain-google-genai
langchain_qdrant
websockets